# docs/swagger.yaml
basePath: /api/v1
definitions:
  comment.commentListResponse:
    properties:
      parent_comment_list:
        items:
          $ref: '#/definitions/model.ParentCommentInfo'
        type: array
      parent_comment_num:
        type: integer
    type: object
  comment.evaluationPublishRequest:
    properties:
      attendance_check_type:
        type: integer
      content:
        type: string
      course_id:
        description: FIX 加上 binding
        type: string
      course_name:
        type: string
      exam_check_type:
        type: integer
      is_anonymous:
        type: boolean
      rate:
        type: integer
      tags:
        items:
          type: integer
        type: array
    required:
    - course_id
    type: object
  comment.evaluationPublishResponse:
    properties:
      evaluation_id:
        type: integer
    type: object
  comment.likeDataRequest:
    properties:
      is_like:
        type: boolean
    type: object
  comment.likeDataResponse:
    properties:
      is_like:
        type: boolean
      like_num:
        type: integer
    type: object
  comment.newCommentRequest:
    properties:
      content:
        type: string
      is_anonymous:
        type: boolean
    type: object
  comment.playgroundResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/model.EvaluationInfo'
        type: array
      sum:
        type: integer
    type: object
  model.ClassInfo:
    properties:
      class_id:
        type: string
      class_name:
        type: string
      course_id:
        type: string
      places:
        items:
          type: string
        type: array
      teacher:
        type: string
      times:
        items:
          $ref: '#/definitions/model.ClassTimeInfo'
        type: array
    type: object
  model.ClassTableInfo:
    properties:
      class_list:
        items:
          $ref: '#/definitions/model.ClassInfo'
        type: array
      class_num:
        type: integer
      table_id:
        type: integer
      table_name:
        type: string
    type: object
  model.ClassTimeInfo:
    properties:
      day:
        type: integer
      duration:
        type: integer
      start:
        type: integer
      week_state:
        description: 全周0,单周1,双周2
        type: integer
      weeks:
        type: string
    type: object
  model.CommentInfo:
    properties:
      content:
        type: string
      id:
        type: integer
      is_anonymous:
        type: boolean
      is_like:
        type: boolean
      like_num:
        type: integer
      target_user_info:
        $ref: '#/definitions/model.UserInfoResponse'
        type: object
      time:
        type: string
      user_info:
        $ref: '#/definitions/model.UserInfoResponse'
        type: object
    type: object
  model.EvaluationInfo:
    properties:
      attendance_check_type:
        type: integer
      comment_num:
        type: integer
      content:
        type: string
      course_id:
        type: string
      course_name:
        type: string
      exam_check_type:
        type: integer
      id:
        type: integer
      is_anonymous:
        type: boolean
      is_like:
        type: boolean
      like_num:
        type: integer
      rate:
        type: integer
      tags:
        items:
          type: integer
        type: array
      teacher:
        type: string
      time:
        type: string
      user_info:
        $ref: '#/definitions/model.UserInfoResponse'
        type: object
    type: object
  model.ParentCommentInfo:
    properties:
      comment_id:
        type: integer
      content:
        type: string
      id:
        type: integer
      is_anonymous:
        type: boolean
      is_like:
        type: boolean
      like_num:
        type: integer
      sub_comments_list:
        items:
          $ref: '#/definitions/model.CommentInfo'
        type: array
      sub_comments_num:
        type: integer
      time:
        type: string
      user_info:
        $ref: '#/definitions/model.UserInfoResponse'
        type: object
    type: object
  model.UserInfoResponse:
    properties:
      avatar:
        type: string
      sid:
        type: string
      username:
        type: string
    type: object
  table.addClassResponseData:
    properties:
      class_info:
        $ref: '#/definitions/model.ClassInfo'
        type: object
      table_id:
        type: integer
    type: object
  table.getTablesResponse:
    properties:
      table_list:
        items:
          $ref: '#/definitions/model.ClassTableInfo'
        type: array
      table_num:
        type: integer
    type: object
  table.renameBodyData:
    properties:
      new_name:
        type: string
    type: object
host: '...'
info:
  contact: {}
  description: 木犀课栈
  license: {}
  title: Muxi K-Stack
  version: "1.0"
paths:
  /comment/{id}/:
    post:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 评论id
        in: path
        name: id
        required: true
        type: string
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/comment.newCommentRequest'
          type: object
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommentInfo'
      summary: 回复评论
      tags:
      - comment
  /comment/{id}/like/:
    put:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 点赞评论id
        in: path
        name: id
        required: true
        type: string
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/comment.likeDataRequest'
          type: object
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.likeDataResponse'
      summary: 评论点赞/取消点赞
      tags:
      - comment
  /evaluation/:
    post:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/comment.evaluationPublishRequest'
          type: object
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.evaluationPublishResponse'
      summary: 发布评课
      tags:
      - comment
  /evaluation/{id}/:
    delete:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 评课id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: 删除评课
      tags:
      - comment
    get:
      parameters:
      - description: 游客登录则不需要此字段或为空
        in: header
        name: token
        type: string
      - description: 评课id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.EvaluationInfo'
      summary: 获取评课详情
      tags:
      - comment
  /evaluation/{id}/comment/:
    post:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 评课id
        in: path
        name: id
        required: true
        type: string
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/comment.newCommentRequest'
          type: object
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommentInfo'
      summary: 评论评课
      tags:
      - comment
  /evaluation/{id}/comments/:
    get:
      parameters:
      - description: 游客登录则不需要此字段或为空
        in: header
        name: token
        type: string
      - description: 评课id
        in: path
        name: id
        required: true
        type: string
      - description: 最大的一级评论数量
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 翻页页码，默认为0
        in: query
        name: pageNum
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.commentListResponse'
      summary: 评论点赞/取消点赞
      tags:
      - comment
  /evaluation/{id}/like/:
    put:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 点赞的评课id
        in: path
        name: id
        required: true
        type: string
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/comment.likeDataRequest'
          type: object
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.likeDataResponse'
      summary: 评论点赞/取消点赞
      tags:
      - comment
  /evaluation/list/:
    get:
      parameters:
      - description: 游客登录则不需要此字段或为空
        in: header
        name: token
        type: string
      - description: 最大的一级评论数量
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 上一次请求的最后一个评课的id，若是初始请求则为空或-1
        in: query
        name: lastEvaluationId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.playgroundResponse'
      summary: 评课广场获取评课列表
      tags:
      - comment
  /table/:
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/table.getTablesResponse'
      summary: 获取课表
      tags:
      - table
    post:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 若是创建副本，则为课表副本id，若是添加新课表，则为空
        in: path
        name: id
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ClassTableInfo'
      summary: 新建课表
      tags:
      - table
  /table/{id}/:
    delete:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 课表id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: 删除课表
      tags:
      - table
  /table/{id}/class/:
    delete:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 课表id
        in: path
        name: id
        required: true
        type: string
      - description: 课堂id
        in: query
        name: classId
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: 删除课堂
      tags:
      - table
    post:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 课表id
        in: path
        name: id
        required: true
        type: string
      - description: 课堂id
        in: query
        name: classId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/table.addClassResponseData'
      summary: 收藏的课堂加入课表
      tags:
      - table
  /table/{id}/rename/:
    put:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 课表id
        in: path
        name: id
        required: true
        type: string
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/table.renameBodyData'
          type: object
      responses:
        "200":
          description: OK
      summary: 课表重命名
      tags:
      - table
schemes:
- https
swagger: "2.0"
tags:
- description: 用户相关
  name: user
- description: 消息提醒
  name: message
- description: 课程相关
  name: course
- description: 评课&评论
  name: comment
- description: 排课课表
  name: table
- description: 举报
  name: report